Remove use of .time files
The submake invocations here seem to confuse make on Haiku. I am not
sure why.

This change may create problems in the day-to-day build of nasm, I did
not study the buildsystem closely to understand the consequences. It is
fine for use in the Haiku package since that just runs one clean build.